# src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ScoutOff API",
    "description": "Backend API for ScoutOff — a decentralised football scouting platform built on the Stellar blockchain (Soroban smart contracts).\n",
    "version": "1.0.0",
    "contact": {
      "name": "ScoutOff Engineering"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Stable API (version-less alias)"
    },
    {
      "url": "/api/v1",
      "description": "API v1"
    },
    {
      "url": "/api/v2",
      "description": "API v2 (includes SSE stream endpoint)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Long-lived scout API key (alternative to a Bearer JWT). Restricted keys may only perform operations covered by their granted scopes (#1019). Legacy keys without a scope list are unrestricted."
      }
    },
...
  },
  "paths": {
...
    "/auth/challenge": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "Issue a SEP-10-style challenge transaction XDR for the given Stellar account.",
        "operationId": "getChallenge",
        "description": "Issue a SEP-10-style challenge transaction XDR for the given Stellar\naccount. The client signs it and exchanges it for a token via POST\n/auth/token.",
        "parameters": [
          {
            "name": "account",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Stellar public key (G...) to build the challenge for"
          }
        ],
        "responses": {
          "200": {
            "description": "{ challenge: string, networkPassphrase: string }"
          },
          "400": {
            "description": "{ success: false, error: string } - Missing/invalid account"
          }
        },
        "security": []
      }
    },
...
    "/auth/token": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Exchange a signed challenge transaction for an access + refresh token pair.",
        "operationId": "postToken",
        "description": "Exchange a signed challenge transaction for an access + refresh token\npair. The caller's role is derived from the verified account: configured\nadmin wallets always get `admin`, otherwise the requested `role` (default\n`player`) is used.",
        "requestBody": {
          "required": true,
          "description": "{ transaction: string, role?: 'validator' | 'player' | 'scout' } - Signed challenge XDR",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "transaction": {
                    "type": "string",
                    "minLength": 1
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "player",
                      "scout",
                      "validator",
                      "admin"
                    ]
                  }
                },
                "required": [
                  "transaction"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "{ token, accessToken, refreshToken, account, expiresAt }"
          },
          "400": {
            "description": "{ success: false, error: string } - Invalid body or malformed XDR"
          },
          "401": {
            "description": "{ success: false, error: string } - Invalid signature or expired challenge"
          }
        },
        "security": []
      }
    },
...
    "/players/register": {
      "post": {
        "tags": [
          "players"
        ],
        "summary": "Register a new player profile.",
        "operationId": "registerPlayer",
        "description": "Register a new player profile. `wallet` must match the authenticated\naccount. Metadata is pinned to IPFS on the caller's behalf unless a\npre-pinned `metadataUri` is supplied instead.",
        "requestBody": {
          "required": true,
          "description": "{ wallet: string, position: string, region: string, metadata?: object } | { wallet, position, region, metadataUri: string }",
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "object",
                    "properties": {
                      "wallet": {
                        "type": "string",
                        "minLength": 56,
                        "maxLength": 56
                      },
                      "position": {
                        "type": "string",
                        "minLength": 1
                      },
                      "region": {
                        "type": "string",
                        "minLength": 1
                      },
                      "metadata": {
                        "type": "object",
                        "additionalProperties": {}
                      }
                    },
                    "required": [
                      "wallet",
                      "position",
                      "region",
                      "metadata"
                    ],
                    "additionalProperties": false
                  },
                  {
                    "type": "object",
                    "properties": {
                      "wallet": {
                        "type": "string",
                        "minLength": 56,
                        "maxLength": 56
                      },
                      "position": {
                        "type": "string",
                        "minLength": 1
                      },
                      "region": {
                        "type": "string",
                        "minLength": 1
                      },
                      "metadataUri": {
                        "type": "string",
                        "minLength": 1,
                        "description": "Additional validation is applied beyond this schema — see the route handler."
                      }
                    },
                    "required": [
                      "wallet",
                      "position",
                      "region",
                      "metadataUri"
                    ],
                    "additionalProperties": false
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "{ success: true, data: { playerId, metadataUri, gatewayUrl } }"
          },
          "400": {
            "description": "{ success: false, error: string } - Invalid body"
          },
          "403": {
            "description": "{ success: false, error: string } - wallet does not match authenticated account"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-required-role": [
          "player"
        ]
      }
    },
    "/players/{playerId}": {
      "get": {
        "tags": [
          "players"
        ],
        "summary": "Fetch a single player profile, including a live `offerCount`.",
        "operationId": "getPlayer",
        "description": "Fetch a single player profile, including a live `offerCount`. Deactivated\nprofiles return 404 to non-owners/non-admins. Supports conditional\nrequests via ETag (304 when `If-None-Match` matches). The same ETag is the\nversion token for optimistic concurrency — echo it as `If-Match` on PUT.",
        "parameters": [
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Player's unique identifier (cuid2)"
          }
        ],
        "responses": {
          "200": {
            "description": "{ success: true, data: PlayerDetail }"
          },
          "304": {
            "description": "Not Modified (when If-None-Match matches the current ETag)"
          },
          "400": {
            "description": "{ success: false, error: string } - Invalid playerId"
          },
          "404": {
            "description": "{ success: false, error: string } - Player not found or not visible to caller"
          }
        },
        "security": [
          {},
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "players"
        ],
        "summary": "Update a player's profile metadata.",
        "operationId": "updatePlayer",
        "description": "Update a player's profile metadata. Owner-only. Accepts either a raw\n`metadata` object (pinned to IPFS by the backend) or a pre-pinned\n`metadataUri`. Optimistic concurrency (#1151): the ETag returned by GET\nmust be echoed in the `If-Match` header. A missing header is rejected with\n428, a stale one with 412 (no write happens); `If-Match: *` is the\ndocumented override.",
        "parameters": [
          {
            "name": "playerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Player's unique identifier (cuid2)"
          },
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "ETag from GET /api/players/:playerId (required; \"*\" to override)"
          }
        ],
        "requestBody": {
          "required": true,
          "description": "{ metadata: object } | { metadataUri: string }",
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "object",
                    "properties": {
                      "metadata": {
                        "type": "object",
                        "additionalProperties": {}
                      }
                    },
                    "required": [
                      "metadata"
                    ],
                    "additionalProperties": false
                  },
                  {
                    "type": "object",
                    "properties": {
                      "metadataUri": {
                        "type": "string",
                        "minLength": 1,
                        "description": "Additional validation is applied beyond this schema — see the route handler."
                      }
                    },
                    "required": [
                      "metadataUri"
                    ],
                    "additionalProperties": false
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "{ success: true, data: { metadataUri, playerId } }"
          },
          "400": {
            "description": "{ success: false, error: string } - Invalid playerId or body"
          },
          "403": {
            "description": "{ success: false, error: string } - Not the profile owner"
          },
          "404": {
            "description": "{ success: false, error: string } - Player not found"
          },
          "412": {
            "description": "{ success: false, error: string } - If-Match does not match the current profile version"
          },
          "428": {
            "description": "{ success: false, error: string } - If-Match header required"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-required-role": [
          "player"
        ],
        "x-owner-only": true
      }
    },
...
    "/validators/milestone": {
      "post": {
        "tags": [
          "validators"
        ],
        "summary": "Submit evidence for a player milestone.",
        "operationId": "submitMilestoneEvidence",
        "description": "Submit evidence for a player milestone. `evidenceUri` may be an\n`https://` URL (downloaded and re-pinned to IPFS) or an `ipfs://` CID\n(recorded directly). Invalidates the player's milestone cache on success.\n\nRate-limited per IP, per validator wallet, and per target player_id (#1137).",
        "requestBody": {
          "required": true,
          "description": "{ playerId: string, milestoneType: 'identity'|'performance'|'trial_offer', evidenceUri: string }",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "playerId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "milestoneType": {
                    "type": "string",
                    "enum": [
                      "identity",
                      "performance",
                      "trial_offer"
                    ]
                  },
                  "evidenceUri": {
                    "type": "string",
                    "minLength": 1,
                    "description": "Additional validation is applied beyond this schema — see the route handler."
                  },
                  "notes": {
                    "type": "string",
                    "maxLength": 2000
                  },
                  "validatorComment": {
                    "type": "string",
                    "maxLength": 2000
                  }
                },
                "required": [
                  "playerId",
                  "milestoneType",
                  "evidenceUri"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "{ success: true, data: { evidenceCid: string } }"
          },
          "400": {
            "description": "{ success: false, error: string } - Invalid body"
          },
          "413": {
            "description": "{ success: false, error: string } - Remote evidence file too large"
          },
          "422": {
            "description": "{ success: false, error: string } - Remote evidence has an unsupported content type"
          },
          "429": {
            "description": "{ success: false, error: string } - Per-player rate limit exceeded; Retry-After header set"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-required-role": [
          "validator"
        ]
      }
    },
...
  }
}